requirements.txt
requests
fastapi
openai
uvicorn
python-dotenv